disable duplicate macOS notarization step (#398)
As of electron-builder v24, there is a built-in notarization step for
macOS builds. This notarization step expects a slightly different set of
environment variable values compared with what the existing Notarize.js
script uses, which causes an error.

Disable the built-in notarization step in electron-builder, so we
continue using the existing Notarize.js script for now.